@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
@@ -7,6 +8,76 @@ import '../core/drm.dart';
 import 'native_events.dart';
 import 'native_tracks.dart';
 import 'playback_options.dart';
+
+/// User-facing playback preferences that persist on the controller.
+@immutable
+class ThaPlayerPreferences {
+  static const Object _unset = Object();
+
+  /// Whether data saver mode is enabled.
+  final bool dataSaver;
+
+  /// Manually selected video track id (or null for auto).
+  final String? manualVideoTrackId;
+
+  /// Manually selected audio track id (or null for auto).
+  final String? manualAudioTrackId;
+
+  /// Manually selected subtitle track id (or null for off/auto).
+  final String? manualSubtitleTrackId;
+
+  /// Playback speed (1.0 = normal).
+  final double playbackSpeed;
+
+  const ThaPlayerPreferences({
+    this.dataSaver = false,
+    this.manualVideoTrackId,
+    this.manualAudioTrackId,
+    this.manualSubtitleTrackId,
+    this.playbackSpeed = 1.0,
+  });
+
+  ThaPlayerPreferences copyWith({
+    bool? dataSaver,
+    Object? manualVideoTrackId = _unset,
+    Object? manualAudioTrackId = _unset,
+    Object? manualSubtitleTrackId = _unset,
+    double? playbackSpeed,
+  }) {
+    return ThaPlayerPreferences(
+      dataSaver: dataSaver ?? this.dataSaver,
+      manualVideoTrackId: manualVideoTrackId == _unset
+          ? this.manualVideoTrackId
+          : manualVideoTrackId as String?,
+      manualAudioTrackId: manualAudioTrackId == _unset
+          ? this.manualAudioTrackId
+          : manualAudioTrackId as String?,
+      manualSubtitleTrackId: manualSubtitleTrackId == _unset
+          ? this.manualSubtitleTrackId
+          : manualSubtitleTrackId as String?,
+      playbackSpeed: playbackSpeed ?? this.playbackSpeed,
+    );
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return other is ThaPlayerPreferences &&
+        other.dataSaver == dataSaver &&
+        other.manualVideoTrackId == manualVideoTrackId &&
+        other.manualAudioTrackId == manualAudioTrackId &&
+        other.manualSubtitleTrackId == manualSubtitleTrackId &&
+        other.playbackSpeed == playbackSpeed;
+  }
+
+  @override
+  int get hashCode => Object.hash(
+    dataSaver,
+    manualVideoTrackId,
+    manualAudioTrackId,
+    manualSubtitleTrackId,
+    playbackSpeed,
+  );
+}
 
 /// Bridge between Flutter and the native player implementation.
 ///
@@ -17,14 +88,31 @@ class ThaNativePlayerController {
   final bool autoPlay;
   final bool loop;
   final ThaPlaybackOptions playbackOptions;
+
+  /// Controller-wide preferences that persist across fullscreen swaps.
+  final ValueNotifier<ThaPlayerPreferences> preferences;
+
+  /// Shared BoxFit state for inline and fullscreen views.
+  final ValueNotifier<BoxFit> boxFitNotifier;
   static int _nextControllerId = 1;
   final int _controllerId;
   MethodChannel? _channel;
   ThaNativeEvents? _events;
+  final ThaPlayerPreferences _initialPreferences;
+  final ValueNotifier<ThaPlaybackState> _playbackState = ValueNotifier(
+    const ThaPlaybackState(
+      position: Duration.zero,
+      duration: Duration.zero,
+      isPlaying: false,
+      isBuffering: true,
+    ),
+  );
+  final ValueNotifier<ThaPlayerError?> _errorDetails = ValueNotifier(null);
+  bool _boxFitUserSet = false;
   int _resumePositionMs = 0;
   bool _wasPlaying = false;
   VoidCallback? _eventsListener;
-  bool _dataSaver = false;
+  VoidCallback? _errorListener;
 
   /// Create a controller with a single media item.
   ThaNativePlayerController.single(
@@ -32,7 +120,12 @@ class ThaNativePlayerController {
     this.autoPlay = true,
     this.loop = false,
     this.playbackOptions = const ThaPlaybackOptions(),
+    ThaPlayerPreferences initialPreferences = const ThaPlayerPreferences(),
+    BoxFit initialBoxFit = BoxFit.contain,
   }) : playlist = [source],
+       preferences = ValueNotifier(initialPreferences),
+       boxFitNotifier = ValueNotifier(initialBoxFit),
+       _initialPreferences = initialPreferences,
        _controllerId = _nextControllerId++ {
     _wasPlaying = autoPlay;
   }
@@ -43,7 +136,12 @@ class ThaNativePlayerController {
     this.autoPlay = true,
     this.loop = false,
     this.playbackOptions = const ThaPlaybackOptions(),
-  }) : _controllerId = _nextControllerId++ {
+    ThaPlayerPreferences initialPreferences = const ThaPlayerPreferences(),
+    BoxFit initialBoxFit = BoxFit.contain,
+  }) : preferences = ValueNotifier(initialPreferences),
+       boxFitNotifier = ValueNotifier(initialBoxFit),
+       _initialPreferences = initialPreferences,
+       _controllerId = _nextControllerId++ {
     _wasPlaying = autoPlay;
   }
 
@@ -53,37 +151,55 @@ class ThaNativePlayerController {
     'loop': loop,
     'startPositionMs': _resumePositionMs,
     'startAutoPlay': _wasPlaying,
-    'dataSaver': _dataSaver,
+    'dataSaver': preferences.value.dataSaver,
     'playbackOptions': playbackOptions.toMap(),
     'controllerId': _controllerId,
-    'playlist':
-        playlist
-            .map(
-              (s) => {
-                'url': s.url,
-                'headers': s.headers ?? {},
-                'isLive': s.isLive,
-                'drm': _drmToMap(s.drm),
-              },
-            )
-            .toList(),
+    'playlist': playlist
+        .map(
+          (s) => {
+            'url': s.url,
+            'headers': s.headers ?? {},
+            'isLive': s.isLive,
+            'drm': _drmToMap(s.drm),
+          },
+        )
+        .toList(),
   };
 
   /// Bind the controller to a platform view id.
   void attachViewId(int id) {
     _channel = MethodChannel('thaplayer/view_$id');
+    // Capture last known state before we dispose the old stream.
+    final previousState = _events?.state.value;
+    if (previousState != null) {
+      _resumePositionMs = previousState.position.inMilliseconds;
+      _wasPlaying = previousState.isPlaying;
+    }
+    if (_eventsListener != null) {
+      _events?.state.removeListener(_eventsListener!);
+    }
+    if (_errorListener != null) {
+      _events?.errorDetails.removeListener(_errorListener!);
+    }
     // Rebind events; dispose previous to avoid leaks
     _events?.dispose();
     final ev = ThaNativeEvents(id);
     ev.start();
     _events = ev;
-    _eventsListener?.call();
     _eventsListener = () {
       final v = ev.state.value;
+      _playbackState.value = v;
       _resumePositionMs = v.position.inMilliseconds;
       _wasPlaying = v.isPlaying;
     };
+    _errorListener = () {
+      _errorDetails.value = ev.errorDetails.value;
+    };
     ev.state.addListener(_eventsListener!);
+    ev.errorDetails.addListener(_errorListener!);
+    _playbackState.value = ev.state.value;
+    _errorDetails.value = ev.errorDetails.value;
+    _applyPreferencesToNative();
   }
 
   Map<String, dynamic> _drmToMap(ThaDrmConfig drm) => {
@@ -106,31 +222,59 @@ class ThaNativePlayerController {
 
   /// Adjust the playback [speed].
   Future<void> setSpeed(double speed) async =>
-      _invoke('setSpeed', {'speed': speed});
+      setPlaybackSpeed(speed, persist: true);
+
+  /// Adjust the playback speed. Set [persist] to false for temporary boosts.
+  Future<void> setPlaybackSpeed(double speed, {bool persist = true}) async {
+    if (persist) {
+      _updatePreferences(preferences.value.copyWith(playbackSpeed: speed));
+    }
+    await _invoke('setSpeed', {'speed': speed});
+  }
 
   /// Toggle looping for the current item or playlist.
   Future<void> setLooping(bool looping) async =>
       _invoke('setLooping', {'loop': looping});
 
   /// Update the content scaling of the platform view.
-  Future<void> setBoxFit(BoxFit fit) async => _invoke('setBoxFit', {
-    'fit': switch (fit) {
-      BoxFit.contain => 'contain',
-      BoxFit.cover => 'cover',
-      BoxFit.fill => 'fill',
-      BoxFit.fitWidth => 'fitWidth',
-      BoxFit.fitHeight => 'fitHeight',
-      BoxFit.none => 'contain',
-      BoxFit.scaleDown => 'contain',
-    },
-  });
+  ///
+  /// When [updatePreference] is true, the value is stored on the controller so
+  /// fullscreen swaps keep the user's choice.
+  Future<void> setBoxFit(BoxFit fit, {bool updatePreference = true}) async {
+    if (updatePreference && boxFitNotifier.value != fit) {
+      boxFitNotifier.value = fit;
+      _boxFitUserSet = true;
+    }
+    await _invoke('setBoxFit', {
+      'fit': switch (fit) {
+        BoxFit.contain => 'contain',
+        BoxFit.cover => 'cover',
+        BoxFit.fill => 'fill',
+        BoxFit.fitWidth => 'fitWidth',
+        BoxFit.fitHeight => 'fitHeight',
+        BoxFit.none => 'contain',
+        BoxFit.scaleDown => 'contain',
+      },
+    });
+  }
+
+  /// Apply an initial fit only when no user choice has been stored yet.
+  ///
+  /// This is used by widgets to keep a controller-wide BoxFit in sync without
+  /// overriding a user-set value.
+  void setInitialBoxFit(BoxFit fit) {
+    if (_boxFitUserSet) return;
+    if (boxFitNotifier.value != fit) {
+      boxFitNotifier.value = fit;
+    }
+  }
 
   /// Re-prepares the media source and attempts to resume playback.
   Future<void> retry() async => _invoke('retry');
 
   /// Caps the bitrate when [enable] is true in order to save data.
   Future<void> setDataSaver(bool enable) async {
-    _dataSaver = enable;
+    _updatePreferences(preferences.value.copyWith(dataSaver: enable));
     await _invoke('setDataSaver', {'enable': enable});
   }
 
@@ -147,11 +291,13 @@ class ThaNativePlayerController {
 
   /// Select a specific video track by [trackId].
   Future<void> selectVideoTrack(String trackId) async {
+    _updatePreferences(preferences.value.copyWith(manualVideoTrackId: trackId));
     await _invoke('setVideoTrack', {'id': trackId});
   }
 
   /// Clears manual track overrides and returns to automatic selection.
   Future<void> clearVideoTrackSelection() async {
+    _updatePreferences(preferences.value.copyWith(manualVideoTrackId: null));
     await _invoke('setVideoTrack', {'id': null});
   }
 
@@ -168,6 +314,7 @@ class ThaNativePlayerController {
 
   /// Selects an audio track. Passing `null` restores the default selection.
   Future<void> selectAudioTrack(String? trackId) async {
+    _updatePreferences(preferences.value.copyWith(manualAudioTrackId: trackId));
     await _invoke('setAudioTrack', {'id': trackId});
   }
 
@@ -184,6 +331,9 @@ class ThaNativePlayerController {
 
   /// Selects a subtitle track. Passing `null` disables text rendering.
   Future<void> selectSubtitleTrack(String? trackId) async {
+    _updatePreferences(
+      preferences.value.copyWith(manualSubtitleTrackId: trackId),
+    );
     await _invoke('setSubtitleTrack', {'id': trackId});
   }
 
@@ -193,14 +343,55 @@ class ThaNativePlayerController {
     return ok ?? false;
   }
 
+  /// Reset playback preferences to the initial values passed to the controller.
+  ///
+  /// When [applyToNative] is true, the current native player is updated
+  /// immediately.
+  void resetPreferences({bool applyToNative = true}) {
+    _updatePreferences(_initialPreferences);
+    if (applyToNative) {
+      _applyPreferencesToNative();
+    }
+  }
+
   /// Release native resources.
   Future<void> dispose() async {
     await _invoke('dispose');
     _events?.dispose();
+    _playbackState.dispose();
+    _errorDetails.dispose();
+    preferences.dispose();
+    boxFitNotifier.dispose();
   }
 
   /// Playback events emitted by the native layer.
   ThaNativeEvents? get events => _events;
+
+  /// Current playback state exposed as a listenable.
+  ///
+  /// Listen to this instead of reaching into platform events directly.
+  ValueListenable<ThaPlaybackState> get playbackState => _playbackState;
+
+  /// Structured error details emitted by the native layer.
+  ValueListenable<ThaPlayerError?> get errorDetails => _errorDetails;
+
+  void _updatePreferences(ThaPlayerPreferences next) {
+    if (preferences.value == next) return;
+    preferences.value = next;
+  }
+
+  void _applyPreferencesToNative() {
+    final prefs = preferences.value;
+    unawaited(setPlaybackSpeed(prefs.playbackSpeed, persist: false));
+    unawaited(setDataSaver(prefs.dataSaver));
+    if (prefs.manualVideoTrackId == null) {
+      unawaited(clearVideoTrackSelection());
+    } else {
+      unawaited(selectVideoTrack(prefs.manualVideoTrackId!));
+    }
+    unawaited(selectAudioTrack(prefs.manualAudioTrackId));
+    unawaited(selectSubtitleTrack(prefs.manualSubtitleTrackId));
+  }
 
   Future<void> _invoke(String method, [Map<String, dynamic>? args]) async {
     final ch = _channel;
